requirements.txt
attrs==23.1.0
barcode==1.0.3
brother-ql==0.9.4
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fonttools==4.45.1
future==0.18.3
kiwisolver==1.4.5
Levenshtein==0.23.0
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
packbits==0.6
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-Levenshtein==0.23.0
pyusb==1.2.1
rapidfuzz==3.5.2
setuptools==69.0.2
six==1.16.0
